Revert all previous commits until 9a7ac4e7 (not included)
During the refactoring of Dr.Jit to use nanobind rather than pybind11,
some commits were still pushed to a pybind11-based `master` branch.
These commits are all reverted here such than we can cleanly apply all
the changes required by the nanobind refactoring.

+++ drjit/traits.py
@@ -546,7 +546,7 @@ def float_array_t(arg):
         type: Result of the conversion as described above.
     '''
     if not is_array_v(arg):
-        return float
+        return int
 
     size = arg.Type.Size
     if size == 2:
